Add wordpress to the dock services
Defining wordpress docker environment. The composefile is based on a version, I found on line.
The wordpress container is based on the lastest tag, so that the we can update the software easily.
Since the newest Version of Wordpress not working with mysql, I choose mariadb as an DMBS.
The Version is fixed. The Actual content is not tracked by git. It should be backup on another way.